feat: add map loader

+++ _build/Debug.cpp
@@ -18,18 +18,18 @@ void Debug::draw()
 
     DrawText(std::format(
         "Going up: {}",
-        gameInstance->getPlayer()->getState().isAscending).c_str(), position.x,position.y, fontSize,
+        gameInstance->getPlayer()->getState().isAscending).c_str(), (int)position.x, (int)position.y, fontSize,
         (gameInstance->getPlayer()->getState().isAscending ? GREEN : BLACK));
     DrawText(std::format(
         "Going down: {}",
-        gameInstance->getPlayer()->getState().isDescending).c_str(), position.x, position.y + 20, fontSize,
+        gameInstance->getPlayer()->getState().isDescending).c_str(), (int)position.x, (int)position.y + 20, fontSize,
         (gameInstance->getPlayer()->getState().isDescending ? GREEN : BLACK));
 
     if (gameInstance->getPlayer()->getState().FacingDirection.right) {
-        DrawText("Facing direction: right", position.x, position.y + 40, fontSize, BLACK);
+        DrawText("Facing direction: right", (int)position.x, (int)position.y + 40, fontSize, BLACK);
     }
     else {
-        DrawText("Facing direction: left", position.x, position.y + 40, fontSize, BLACK);
+        DrawText("Facing direction: left", (int)position.x, (int)position.y + 40, fontSize, BLACK);
     }
     
 }
